+++ http-netty/src/main/java/io/micronaut/http/netty/PublisherAsStream.java
@@ -17,7 +17,6 @@ package io.micronaut.http.netty;
 
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.annotation.NonNull;
-import io.micronaut.core.async.subscriber.PublisherAsBlocking;
 import io.netty.buffer.ByteBuf;
 
 import java.io.IOException;
